集成mybatis-plus查询结果泛型的封装

+++ flycloud-common/flycloud-common-database/src/main/java/com/fly/common/database/domain/vo/PageVo.java
@@ -20,7 +20,7 @@ public class PageVo<T> implements Serializable {
     /**
      * 总页数
      */
-    private Integer pages;
+    private Long pages;
 
     /**
      * 总条目数
